+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import '../src/app/stylesheets/styles.css';
+import './stylesheets/styles.css';
 import Home from './app/pages/Home/Home';
 
 function App() {
