+++ src/stylesheets/layouts/_banner.scss
