+++ src/stylesheets/modules/_badge.scss
