+++ src/stylesheets/modules/_product.scss
