+++ src/stylesheets/pages/_all.scss
