+++ src/stylesheets/styles.css
@@ -928,6 +928,7 @@ h3 {
   }
   .home-page .section-banner {
     height: 740px;
+    background-image: url(../assets/images/Fashion-home-image-9.png);
   }
   .home-page .section-about {
     margin-bottom: 48px;
@@ -939,6 +940,7 @@ h3 {
     margin-bottom: 36px;
   }
   .home-page .section-contact {
+    background: url(../assets/images/Fashion-home-image-10.png) no-repeat center/cover;
     height: 217px;
     margin-bottom: 0;
   }
